requirements.txt
mesa>=3.2.0
climada-python>=4.1.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.8.0 